requirements.txt
accelerate==0.31.0
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==23.2.0
Brotli==1.0.9
certifi==2024.7.4
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.0
filelock==3.13.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
gritlm==1.0.1
huggingface-hub==0.23.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
mteb==1.12.79
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.0
pandas==2.2.2
peft==0.11.1
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
polars==1.1.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
ray==2.22.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentry-sdk==2.8.0
setproctitle==1.3.3
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
submitit==1.5.1
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
torch_geometric==2.5.3
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.41.0
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.0
wget==3.2
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
